pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-pipeline"
version = "0.1.0"
description = "Real-time stock data pipeline with async processing"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",           # Async HTTP client for API calls
    "sqlalchemy>=2.0.0",        # ORM for database
    "aiosqlite>=0.19.0",        # Async SQLite driver
    "fastapi>=0.104.0",         # REST API framework
    "uvicorn>=0.24.0",          # ASGI server
    "pydantic>=2.0.0",          # Data validation
    "python-dotenv>=1.0.0",     # Environment variables
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
